2018-09-23/src/unicorn/requirements.txt
click==6.7
Flask==1.0.2
gunicorn==19.9.0
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
mysql-connector-python-rf==2.2.2
Werkzeug==0.14.1
